+++ 6. Effect Ruducers Context/src/components/Home/Home.js
@@ -1,16 +1,18 @@
-import React from 'react';
+import React, { useContext } from 'react';
+import AuthContext from '../../auth-context';
 import Button from '../UI/Button/Button';
 
 import Card from '../UI/Card/Card';
 import classes from './Home.module.css';
 
-const Home = (props) => {
-  return (
-    <Card className={classes.home}>
-      <h1>Welcome back!</h1>
-      <Button onClick={props.onLogout}>Logout</Button>
-    </Card>
-  );
+const Home = () => {
+	const ctx = useContext(AuthContext);
+	return (
+		<Card className={classes.home}>
+			<h1>Welcome back!</h1>
+			<Button onClick={ctx.onLogout}>Logout</Button>
+		</Card>
+	);
 };
 
 export default Home;
