+++ 6. Effect Ruducers Context/src/index.js
@@ -1,12 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
+import { AuthContextProvider } from './auth-context';
 import './index.css';
 
-
 ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+	<React.StrictMode>
+		<AuthContextProvider>
+			<App />
+		</AuthContextProvider>
+	</React.StrictMode>,
+	document.getElementById('root')
 );
